requirements.txt
# 核心依赖（锁定大版本，满足下载/数据库/后续 UI 需求）
yt-dlp>=2024.12.13,<2026.0.0
streamlit>=1.36,<2.0
pydub>=0.25,<0.27
openai>=1.54,<2.0
python-dotenv>=1.0,<2.0
openai-whisper>=20231117
tenacity>=8.2,<10.0
# Python 3.13 开始移除内置 audioop；条件性引入 LTS 版本以保持兼容
audioop-lts>=0.2.1; python_version>="3.13"
